myproject/requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary>=2.9
Pillow>=10.4.0
djangorestframework==3.15.2
requests==2.26.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.51.0
# dj-rest-auth==2.1.9 

